rename isExists into isExisting
