+++ src/fp/isExists/index.ts
@@ -1,6 +1,6 @@
 // This file is generated automatically by `scripts/build/fp.ts`. Please, don't change it.
 
-import fn from '../../isExists/index'
+import fn from '../../isExisting/index'
 import convertToFP from '../_lib/convertToFP/index'
 
 export default convertToFP(fn, 3)
