+++ src/index.ts
@@ -109,7 +109,7 @@ export { default as isAfter } from './isAfter/index'
 export { default as isBefore } from './isBefore/index'
 export { default as isDate } from './isDate/index'
 export { default as isEqual } from './isEqual/index'
-export { default as isExists } from './isExists/index'
+export { default as isExisting } from './isExisting/index'
 export { default as isFirstDayOfMonth } from './isFirstDayOfMonth/index'
 export { default as isFriday } from './isFriday/index'
 export { default as isFuture } from './isFuture/index'
